Add feedback generation methods and restructure handleCommand() code
The feedback displayed can be decorated to look better. Feedback
generation should be done in this class. The creation of a task can also
be done in the agent as command should not have task creation ability.

In light of this, let's create feedback generation methods for each of
the creation/update/retrieval. Make sure the taskList is updated within
the handleCommand() method and respective feedback are produced for each
command request.