Add greet, echo and exit functionality
Duke now greets the user, will echo the user's input, and the program
will exit when the user inputs "bye". Duke's name has also been
changed to "Star Bot", as in that is the name that it goes by during
user interaction, but the class and file name is left as "Duke" for the
purposes of this assignment.