Resolve sad face symbol's imcompatability with GBK encoding
The sad face in the error handling messge cause compile issue as the
terminal cannot resolve the symbol. There are some leftover "duke."
prefix in EXPECTED.TXT which causes test failure.

Let's remove the sad faces in the error message for now and clean up
the EXPECTED.TXT, and enforce file encoding in runtest.bat to UTF-8.