Duke class: refactor route function for less repetition
There is a lot of repeated logic for parsing and validating input in
the route function.

Let's:
- extract the common logic into helper functions
- create a new type of exception (DukeRoutingException)

Exceptions are thrown by the helper functions to make it easy to break
out of the control flow at many different points within the route function.

The exceptions are caught within the route function, rather than
becoming a checked exception for the route function, so as to keep the
run function clean.