Feat: Implement add, list features (Level-2)
firstly, created a constructor for Duke to make it easier to hold
state/variables. Also made some of the utility functions non-static.

added the `add` and `list` features. Uses an arraylist to store and
display strings input by the user. Created the functions `addItem`
and `printList` to carry this out.

minor fix: the indent function was not working properly for strings
with multiple lines.

note: will start using conventional commits style messages for
future commits.