Feat: Implement tasks+mark as done (Level-3)
Added a new class `task` to keep track of tasks. Also replaced the
ArrayList<String> items with an ArrayList<Task>. Task can be pretty
printed with toString.

Added mark as done feature: Duke will look for strings beginning
with `done ` and parse the rest of the string into an int idx.
The task with index idx will be marked as done.
TODO: add error handling for strings with wrong formatting.