Add subclasses for tasks
A task only has a description and isDone fields. For tasks that
are required to be done before a certain date/time and tasks
that start at a specific date/time, this is insufficient.

Exploiting polymorphism enables the chatbot to display the
additional information with liitle modification.

Let's add subclasses of task to include the date/time field
and to distinguish between the different types of tasks.