Change isExit() method in Command to static method in ExitCommand
The isExit() method examines the content of the command, which will
does not use a getter. I consider the check of Exit command more natural
when it is a method always associated with the ExitCommand Class.

Hence, let's remove the isExit() command in Command class and replace
its functionality with a static method isExitCommand() in ExitCommand
class. Replace the exit condition in do-while loop in Duke class with
the isExitCommand check as well.