Modify the Task class to an abstract class
The Task class is separated into 3 different types, a user input should
always specify one of the 3 types. Hence it is unlikely to create a new
Task object directly.

In light of this, let's make the Task class an abstract class, change
the complete() method in Task class to abstract method as well.