Refactor: Big refactoring of project
1. Created a TaskList class to manage task list related operations.
Moved the `markItem` and `deleteItem` methods from Duke there.

2. Created the Command enum and CommandExecutable functional
interface. each command will have a `match` and `dispatch` method
to respectively handle matching and executing the command.
There are many other nontrivial details that will be elaborated
on in the docs (coming soon).

3. Created the parser class to handle parsing/matching of the
input strings from the scanner and the commands.

4. Moved small misc functions out to Util class. These functions deal
mainly with reusable printing operations. Will change this to a full
blown UI class in the future.