Chatbot.java: standardize print statements
Print statements are found across multiple methods, with some
relying on printHeader() & printFooter() and others on
printWithDecorations(). This is because printWithDecorations()
only indent the text once, which is not visually pleasing
with multi-line text.

As the print statements are so scattered, it becomes difficult to
modify them, i.e. need to edit several lines to achieve consistency.

Let's allow printWithDecorations() to indent once for every newline
character in the text and standardize print statements to use
printWithDecorations().