Add the skeletal version of Duke
Let's start by greeting the user. Duke will echo commands entered by the user, and exits
when the user types the keyword "bye".

To enhance user experience, reserved keywords are added which the user can type
"help" to view all of those, and there are also visual lining breaks to provide a
clearer view of events.