Add support for tracking 3 types of tasks
Let's add 3 new conditions to handle 3 special input types of tasks.
This will allow the user to see which types of specific tasks there is
remaining in the list and eases task handling for them.

We should also let user know the number of pending tasks there is
remaining.

Let's,
- add a line of output that allow users to know the number of tasks
  they have remaining after they added a new task.