Add support for three types of Task in Dino
Let's implement 3 subclasses of Task, namely:
Todo, Deadline and Event. Todos are tasks without
any date/time, Deadlines are tasks that need to be
done before a specific date/time, and Events are
tasks that start at a specific time and ends at a specific time.

Inheritance is used to support multiple task types,
as they are all tasks and there is similarity between them.