Create a Storage class to manage the task data I/O
The Storage class needs to handle save and load operations on data.
When the file/folder to store the task data are not present, the class
also need to detect them and create the correct file/folder immediately
upon loading of data file.

In light of this, let's:
* build a constructor for Storage class to store the file's folder path
and file path in relative path format
* create a createPath() method to create new path for data storage
if current path specified is not detected
* create save() and load() methods for data file reading/writing