Removal of unused files and renaming of CommandReader.java
After reading the project requirement, it becomes clear that DukeBuffer
are now simply the TaskList, hence it provides little utility. The
ErrorCommand class can be removed as the error handling in the code
becomes more fluent. The CommandReader class should be renamed as Parser
as the level-9 requirement suggests.

In light of this, let's remove the 2 files and rename the CommandReader
class into Parser class, with its parse() method turned into static to
make Parser a utility class.