Current packages maintenance and JavaDoc supplementation
The current packages are rather complete. Some minor issues like the
Event and Deadline constructor with schedule parameter in the string
representation can be removed as LocalDate type can be created when
the task is created. Some public method's JavaDoc can also be added.

Let's drop the constructor mentioned above, and add in a few more
JavaDocs for public methods.