Adds todo, deadline and event class
Todo deadline and event class are subclasses of Task.
Todo class is for tasks with no date/time attached
Deadline class is for tasks with a data/time to indicate when the
dealine is.
Event class is for tasks with date/time to indicate when it starts and
when it ends.

Data/time are strings.

The main includes todo, deadline and event features. If user specifies
which type of task to do, the main will handle each type accordingly.