Modify constructor of Task and subclasses
Constructor of Task and subclasses Todo, Deadline and Event accept
one argument for its description, and the isDone field is set to
false.

To load task list data from disk, we must be able to create tasks
with isDone field set to true.

Let's add an isDone parameter to the constructor of Task and its
subclasses, which will allow us to dictate the isDone field when
we create new task objects.