Add ability to parse date and time
Program saves event and deadline task's date and time as a string. This
prevents the program from performing more advanced operations on the
tasks such as searching by date.

Let's allow the program to parse and store the date and time of events
and deadlines using the yyyy-MM-dd[ HH[:]mm] format, where the square
brackets represent optional parameters.