Refactor Duke.java to improve readability and extensibility:
Original code contained UI and logic in the main method.

As a result, if there are any changes to be made to a specific scenario, problems are not isolated into individual methods or cases.

Simplified the driver method to be solely responsible for console input and dependent on other static methods.

By separating the responsibility into smaller chunks, it will improve debugging process if necessary.

Java Doc comments have been added to describe individual methods as well.