Add Date and Time functionality
Duke reads all date and time as String object.

Reading date and time as String limits the things Duke can do. Comparison of dates and times cannot be made as it is a String object.

Let's change the String object to LocalDate and LocalTime object instead.

In this way, Duke would be able to compare the Dates and organise tasks with respect to the dates on command.