Add save functionality
Duke is not able to save task lists upon exiting the program.

Users thus cannot keep a saved task list unless they do not
terminate the program.

Add save functionality in which Duke writes the task list to a file on
the hard disk whenever the task list changes. Also handle cases when
the path to the save file does not exist - save file or directories in
path to save file does not exist.

Saving to the hard disk this way follows the Level 7 requirements.