Make Task an abstract class and add hasDateTime method
Task is a normal class.

There are some methods like hasDateTime that are implemented
differently depending on the subclass of Task.

Let's make Task an abstract class to provide this common
functionality across the subclasses while also allowing Task to
have default methods.

Using an abstract class is preferable over an interface because the
subclasses have an "is-a" relationship to Task, instead of a
"can-do" relationship.