Change Task subclass constructor arguments
Subclasses of Task (Deadline, ToDo, Event) take
in user command Strings as arguments which are
then parsed in the class. However, user commands
are not very relevant to these classes.

Let's change the arguments to more relevant
fields. Such as 'deadline' and 'task' for Deadline.
The parsing of commands into these fields
can be done in the Parser class