Refactor code in Duke.java
Duke.java is handling many different responsibilities - including user
interaction, loading and saving files, and parsing and executing
commands.

This is not in line with the single responsibility principle that
states that a class should only manage a single part of a program's
functionality.

Let's refactor the code in Duke.java to:
* Ui.java to deal with interactions with the user
* Storage.java to deal with loading tasks from and saving tasks to the
save file
* Parser.java to deal with making sense of user commands
* TaskList.java to deal with manipulating the task list
* Command.java and its subclasses to define and execute user commands

This follows the single responsibility principle by creating classes
as necessary that each manage a single part of the program's
functionality.