Organize classes into packages
All the java files are in one single directory.

This makes navigating the directory to access certain classes difficult
as there are many files to sieve through. It also does not provide
encapsulation that packages provide and does not make it clear upon
looking through the directory what types of classes there are - classes
that deal with user commands, classes that deal with tasks, and classes
that deal with excpetions.

Let's organize the classes into suitable java pacakages:
* duke - as the highest level package containing all other packages /
  java files
* duke.command - containing all classes that deal with user commands
* duke.exception - containing all duke excpetion classes
* duke.task - containing all classes that deal with tasks

This makes the code base neater and hence easier to navigate and
understand, and also allows for encapsulation via packages.