Deadline, Event classes: store date using LocalDate class
The deadlines and events' time are being stored using string.

Unable to use the time in a more meaningful way (e.g finding events/
deadlines that is happening on a specific date).

Store time in LocalDate objects whenever possible.

Make it possible to sort tasks by date or look up tasks occuring in a
specific date.