Implement level-7
The past tasks added or deleted are not saved in a file before.

It should be saved in a file for retrieval and the file should be updated after each operation, so that data would be remembered after the operations.

Let's create a Storage class to store the data related to tasks and enable write to the file and read data from the file.

This is done so that the user could have easy reference to tasks they added before.