Add JUnit test for some Ui and TaskList functions
Viscount is not undergoing any unit testing.

It is difficult to confirm that Viscount is working as intended with
only text UI testing. It is also beneficial to test each part of the
program separately.

Let's add JUnit tests for some of the functions in Ui and TaskList
to confirm that these parts of Viscount are working as intended.