Add Level-9 code
For level 9, I have added a FindCommand class in Commands to complete the new find task.

Then, there is another FindException class to handle the case where the keyword mentioned is absent or the keyword is absent.

If there are no exceptions, then different tasks in the currents tasks in taskList is being displayed.

Then, the codes just typed in follows the coding style mentioned.