Add ability to save tasks upon command
Tasks were saved only upon giving an ExitCommand.
This meant that tasks would not be saved if user quits the program
abruptly. We should not expect the user to quit the program properly
everytime.
Let's allow the program to save the tasklist to the hard disk whenever a
task is changed, added, or deleted.