Add task finding functionality
Duke cannot find tasks in the task list based on keywords.

This makes it inconvenient for users who want to do so, which is a
common feature for a task list.

Let's add task finding functionality in which users can specify
keywords and Duke will then search through the task list and return
tasks that contain those keywords, if any.