Add find functionality to list command
Users can list their tasks and can filter by task type or by date
and time of task.

Users should also be able to filter their tasks by the description.

Let's add a find functionaly to the list command to allow users to
filter by description in conjunction with the rest of the list
modifiers.