Adds Save function
Part of the project requirements.


Save the tasks in the hard disk automatically 
whenever the task list changes. 

Load the data from the hard disk when Duke starts up.

Data loaded in from "/data/duke.txt".
May not be OS independent.

Loading of data is not equipped to handle data not in
specific format.