Update Level-7
At Level-6, Duke has all the basic functions of a reminder.

However, Duke cannot store all information into the hardware, which makes it meaningless if it functions as a reminder.

Update is made to include the use of hardware. Now, with the use of adding event, list, done, delete, corresponding information in ./data/Duke.txt will also change accordingly.