Add Find Feature
Add find function in task list

Tasks can be viewed on through the "list" command, which can get more difficult to find the relevent task as the list size increases.

Now users can search for a task using a keyword with the help of Find command:
 - find <keyword>
This will show a list of tasks that contains the key word given.