Adds Date recognition for Deadline Tasks
Part of Project Requirements (Level-8).

Takes in dates for Deadline within strings in the 
format of "yyyy-mm-dd" and displays it as "dd MMM yyyy".

Some code was added to Event class to prepare for
future modifications that will allow Duke to recognize
both Date and Time within the command string of an
Event Task.