Add Level 7 save feature
Any changes made to the list of tasks does not get saved, and every run
of the program would give an empty list at the start, which is not
useful to the user.

Adding save feature would allow the program to save the changes to
the list of tasks from the previous run of the program.

Let's
* save the list of tasks in a duke.txt file under data folder
  every time there are changes to the list of tasks
* read data from the duke.txt file at the start of the program
  run so that the data from previous program can be called and
  used in the program