Update Level-8
For Level-7 and levels before, event and deadline tasks need to specify the respective time.

However, this time can be in any words, which sometimes does not even make sense.

In this level, the format of time has been formated and the user must follow the time format in order to note down the task.