Add search functionality for duke
Users can now find a task by searching for a keyword.

Previously, users have to go through the whole list of task to find what they are looking for which can be inefficient.

Let's add a search feature for users to find the tasks they want by searching for a keyword.