Add Save feature for Level 7
Tasks which are stored in the list while the program is run
is not saved and stored when the user exits the program.

Enabling a save function would improve the usability of this
program as tasks which were entered previously can be
remembered.

Let's providing a saving function which stores the Task list into
a text file when the user exits and retrieves this list when the
program is started up again.