Fix issue with text file saving.
Working directory is different when building and running Duke in the IntelliJ IDE
compared to when running runtest.bat.

Modifications have to be made to the filepath specified depending on which
is the working directory.

Let's use a conditional expression to set the correct file path depending on what
was the original working directory.