Refactor code in Parser, Executor, Controller and Command modules
Reason: Executor class is returning a parsed response to the user. It
should focus on executing the logical commands on the TaskList only.

Controller now explicitly stringifies a response based on each command
and returns it to the Ui. The response is a reply that shows what was
done by the Executor in carrying out the command.

Command no longer has methods to create different kinds of ADD commands:
EVENT, DEADLINE and TODO. Instead all can be grouped under ADD since the
response is the same: adding to the TaskList.

Nit: Fix style in Parser and added comments.