Add level-7 Save implementation and Storage.java
Data is not persistent and flushes on each run.

This is not ideal for long term usage of this task manager.

Let's make the data persistent by saving and loading the tasks to and
from a "saved-tasks.txt" file. Let's also introduce OOP practices by
creating a Storage class to handle this.