Refactor code to extract out Ui, Storage, Parser, TaskList, Command classes
Classes have closely related code can be extracted into a single class.
It reduces the complexity of viewing things.

Ui class: deals with interactions with the user.

Storage class: deals with loading tasks from the file and saving tasks
in the file.

Parser class: deals with making sense of the user command.

TaskList class: contains the task list. It has operations to manipulate
tasks in the list and get data from it.

Command class: abstract Command class to allow us to use polymorphism
when dealing with different commands