Automate project builds using Gradle
Building the from source code and running testcases can be a and be tedious multi-step process.

Gradle integration saves precious development time and labour.

Gradle is being used to automate building and running of Duke. It is also set up to run the whole test suite of unit tests.

This saves the number of clicks to run duke and unit tests.

Use Gradle to automate some of the build tasks of the project.