Use checkStyle to detect coding style violations
Checking for coding style is a manual process which is easy to miss by the developer and takes up time when it can be automated.

Automate the process of coding style checks using Gradle by adding checkstyle.xml with coding style specifications and suppressions for exceptions to ignore.

This provide specification for the checkstyle plugin to follow.