Fix OOP Extensions
More Classes have been added to decentralise the logic,
such as the Storage, Parser, and Ui Class.

However, the classes were not feature complete, especially
the Ui Class.

Let's complete these classes, revamp the logic in the main
Duke class to use these classes, and test it out with text-ui-test.