Package Classes into Duke and Task Packages
The numerous classes in the java directory has become
messy as the number of classes increased.

To improve the organization of the code base, packaging ought
to be done.

Let's package the classes into two main packages - task for the
Task Classes and duke for all the other classes. More packages
will be created in the future if complexity increases. Access modifiers
have been made package-private in the duke class for tighter
security.

Furthermore, minor edits to the code have been made such as
improved exception handling, and fully delegating console output
to the Ui class.