Add Ui, Storage, Parser and TaskList classes
The List of classess is as follows:
1. Ui: Deals with interactions with user and manages System output

2. Storage: Deals with loading tasks from duke.txt and writing new
data to the file.

3. Parser: Deals with intepreting user inputs and returning the
parsing it as their respective tasks or modes. enum Modes is used to
differentiate between each command to tell the program what action
is required.

4. TaskList: Encapsulates the list of tasks as an Array List and
contains methods such as add/remove/markDone to modify the queue.

The Task.java file contains the task class and its subclasses.
DukeException file contains the DukeException class.

The main method in Duke.java only contains the logic of the program.