Add a GUI to Duke
GUI is running solely in the command line with no graphical user interface.

A purely CLI interface may not be as user-friendly as a GUI interface which
more users would be accustomed to.

Use the JavaFX technology to implement the GUI. View of the application is
rendered using FXML which is manipulated using by Controller classes
MainWindow and DialogBox.

To add sub nodes individually to the main node for building the view is
harder to visualize and makes code more verbose.