Add Find Command (Level 9)
Find Command skeleton has been part of Duke so far
but the logic has not been incorporated and tested.

Find command is useful for users to search any tasks
by their keyword(s).

To improve the usability of this program, let's provide
the logic of the find command, including testing via
text-ui-test.