Review code and rectify coding quality violations
Some parts of current code do not comply with coding quality constraint
such as "Magic Number" avoidance, naming and nested conditioning.

These violations should be corrected to make the overall code more
robust and informative.

In light of this, let's:
* rename those inappropriate methods and variables
* bring the conditioning in if-else clause out, thus creating 2
conditioning clauses on the same level
* change request keyword into static variables attached to each command
type
* extract datetime formatter out as a standard pattern for task classes
* rename the format() method to avoid confusion with the Datetime's
in-built format method.