Add  javadocs for new methods and classes and ran checkstyle
Previous commit did not include javadocs and some of the code do not adhere to the coding standard.
For documentation purposes, include javadocs for new methods and classes and update them for the other methods and classes.
To ensure that the coding standard was adhered to, run and pass checkstyle.