Add new functionality: reminder command
The command asks for user input in the form of "remind + number" where
the number is the number of tasks not done and whose schedule are most
recent. The tasks displayed are sorted in their schedule order, with
the tasks whose schedule has past the day getting indicated with an
expiry sign.