Improve find command
The Find command can only filter by keywords found in the task string.
It cannot filter by type or status.

This makes it harder for users to filter by type or status as they will
have to search for keywords such as [T].

Let's improve the find command to make it support more keywords to
filter the search by.