Remove dead codes, magic numbers in Parser, Add throws for exceptions, Get static task deliminter from 1 source.
Removed dead codes from Duke.java. The old run method is not useful for GUI.
Parser to change the substring numbers to commandString and contentString, extract the multiple use of strip.
Added throws DukeException for those method that does not have it in the method signature.
Task_Delimiter will get from Util class, standardize for the whole programme. (single point to change delimiter)