Fix CLI cross and tick not appearing correctly
Duke's CLI version does not display the crosses and ticks of tasks
correctly ever since the GUI version was implemented.

This makes the CLI version not suitable for use.

To continue to support the CLI version of Duke, separate methods for
printing the correct crosses and ticks have been added.