Feat: Add assertions
Add several assertions in the code. Note that most of the code does
not have many assumptions (apart from variables being non-null), and
the only place I've found they're relevant is in the Ui classes.

More non-null assertions may be useful when the variable passed is
not being used immediately, e.g. in constructors.