Add UndoableCommand interface.
Command which cannot be undone have to implement the undo method.

Let's add in an interface that extends the Command interface.

This allows only Command which can be undone to implement the undo method. Command which cannot be undone would not need to know the existence of such undo method.