Implement more flexible command syntax
There is only 1 valid command keyword per command and it is
case-sensitive.

Users will have to enter the full word and in the correct casing in
order for it to be recognised as a valid command keyword. This is an
unnecessary restriction.

Let's add a list of valid command keywords to each command such that
users can invoke a command via a number of different sensible keywords.
Also, let's make command keywords case-insensitive.

In this way, users have more flexibility in invoking commands.