Add help command
PandaBot does not have any means to allow users to be sure that they are giving the correct input format.

This could confuse users who are not familiar with PandaBot's features.

Add a help command which will bring up a help page that contains a list of commands that the
user can use.

Let's make this help page to help familiarise users with PandaBot's available commands.