Shift parsing functionalities to Parser class
The command classes currently perform additional parsing of user
input to retrieve the necessary information for command execution.

This is not the ideal implementation since all parsing of user inputs
should be handled by the Parser class.

Let's abstract all parsing methods in the Command classes to the Parse
class.