Handle error message better.
Error messages are created at the place they were found, e.g.
in parser, storage.

This should be the job of another class, and does not allow
for scalability and ease of change.

Let's pull out all the error messages, and create a new class to
handle them.

The use of enumerations also improves code readability and quality.