Improve Code Quality
Duke's logic has been centred in the Duke class thus far.

However, the code quality is not very good as the getResponse
method is very long and does too many things.

In order to improve the quality of the code, let's abstract the logic
into 2 areas: 1) Parser class will process the user's input and
return the right Command. 2) Introduce Command classes which
can be executed to return a String response. Thus, the logic in Duke
would be abstracted out into Commands which are executed until
the command signals duke to stop running with the continueDuke()
method.