Add BCD Extension C-Help, Fix Bugs
Users might be lost as to how to use Duke as there
is no guide provided to them.

Let's introduce a help command, which produces
useful instructions on what each command does
and how to use it. A help message is also produced
at the start of Duke.

Bugs were also fixed: Exception is now handled
when non-numbers are input for the delete and
done commands.