Rework Storage Class
Storage has depended on Duke supplying a
correct file path of where the saved tasks were
stored.

However, Duke should not need to know the file
path and this function should be delegated to the
Storage class.

To improve the abstraction of Storage's function,
let's shift the file path to within the Storage class
such that everything related to the file I/O is
contained within the Storage class.