Update help command
Help command displays a list of all commands.

This means that every time the user guide is updated, the information
displayed by the help command needs to be updated as well. The
information displayed on the app is also limited as compared to the
user guide.

Let's redirect the user to the user guide when the help command is used.